ALLJOYN-909: Insure that GUID in found name is correct under case where we are redirected and assume the minion role.
When redirected we don't have the GUID information, only the BDADDR
and PSM.  If we retain the master role then the GUID is available in
the node start parameter of the SetState method reply.  If we assume
the minion role, then the GUID must be gotten from the found nodes
parameter instead.