Disable incremental linking for arm build as a workaround to a arm complier bug.
http://stackoverflow.com/questions/11478055/lnk2013-error-fixup-overflow

"This is due to an unfortunate bug in incremental linking on ARM, and can
be worked around by disabling incremental link. As mentioned in the other
answer, this is because the called function and callsite are so far away
from each other that the linker needs to insert long branch islands (and
there is a very particular bug with creation of long branch islands when
using incremental link).

To my knowledge, this bug was not fixed for RTM (when I discovered this
issue myself it was too late to fix for the RTM release) but it is known
and planned to be fixed for a future release."