Updated the Visual Studio build files for the C++ samples
The Visual Studio build files have not been updated since before the bundled daemon
was set as the default way to build windows projects.  The samples would build but
they would always fail saying the could not find the transport because the "null:"
transport was not being built into the projects.  I also modified the way the
libraries are imported when using the ALLJOYN_SDK_HOME environment variable so it
reflects the layout of the distributions that we currently have in the AllJoyn SDKs
posted to alljoyn.org