Doxygen should not be run on the alljoyn_c header files
When doxygen is producing documentation for the C++ API it should only include
the header files from alljoyn and qcc.  The documentation was also processing
the header files for alljoyn_c.  The list of files used for documentation has
been updated so that the alljoyn_c header files are excluded.

Change-Id: Ib57d15d87ec9f11b966a77ac4484670ea26bcea3