New test added to ProxyBusObjectTest
The SecureConnection test sets up two Buses with Peer Authentication
enabled.  It then calls ProxyBusObject.SecureConnection();  This will
cause Authentication to run.

This is copying a similar test found in the JS bindings.  It was added
here to check if some unusual errors seen in the JS were also seen in
C++.

Change-Id: Ia9e1f6263d936a0c1e83a84f027f5085733a57f0