Added two additional tests to SessionTest
BindMemberAddedRemoved a test added while debuging JavaScript unit tests
The test creates a multi-point session and adds and removes BusAttachments
to that session and checks that SessionMemberAdded and SessionMemberRemoved
callbacks are called when they should be called.

Added SessionJoined test.  The test is currently dissabled it is there
to show reproduction of a bug found while working on the BindMemberAddedRemoved
test.  See ALLJOYN-1602 for more information.

Change-Id: I0601660ba4c7af7e36f7b1616ab3ee442719e677