ALLJOYN-1595 test reproducing error seen in JavaScript bindings
The test ProxyBusObjectTest.testSecureConnection kept on failing
with the status ER_BUS_ENDPOINT_CLOSING.  The C++ unit test
ProxyBusObjectTest.SecureConnection was originally developed to see
it the same issue could be reproduced.  Upon further investigation
it was discovered that the test needed to be done using the Async
version of the SecureConnection code.  This test does the exact same
thing as the SecureConnection however using SecureConnectionAsync.

Change-Id: Ie12bb6f99f73506d4a67c5228b88e2024a454090