ALLJOYN-1719 WinRT Alljoyn application crashes at ~MsgArg()
In WinRT binding, when an ajn::MsgArg array is allocated on heap, the
pointer p is stored in a list of type void*. In the destructor, the
operation to release the allocated memory "delete []p" causes assertion
failure "_ASSERTE(_BLOCK_TYPE_IS_VALID(pHead->nBlockUse));". That is
because p is of type void*, and the behavior of "delete []p" is undefined.
The fix is to store the pointer p explicitly as type ajn::MsgArg.