ALLJOYN-1676 scons incorrectly assumed daemon_libs[1] was a *.lib file
The scripts used the fact that the libraries were always in the same location
when in the daemon_libs variable. With the change in build_core for ALLJOYN-1676
this was no longer always true. So we must search the daemon_libs variable list
to find the scons.file.nodes that contain the shared and static daemonlib libraries.