ALLJOYN-1908 Fixed path parsing error in ProxyBusObject
Fixed path parsing error that was in the following functions
 - GetChild
 - GetManagedChild
 - AddChild
 - RemoveChild

The error caused the last letter of a path to be cut off when creating or
reading a child child ProxyBusObject.  This made it so two paths that differed
by a single letter at the end of the path name would look the same.

Change-Id: I8e72b90140f4e220d1e603360aafa5e13d560603