ALLJOYN-1654 Debug levels are inappropriate for some messages
The general approach taken was if the error is a result of shutting down the
bundled daemon then ENDPOINT_CLOSING, and BUS_STOPPING errors are expected and
should not create LogErrors that show up on the users screen. In some cases
where it could be useful for debugging the logErrors were changed to
DbgHLPrintfs in others nothing is printed even in debug mode.

Other code that would regularly print to the screen was modified to print to the
high level debug.  Things like "AllJoyn Daemon GUID" and "Using BundledDaemon"

Change-Id: I4aa2c583beb4ad5cedf2c371a450241711a03440